pyproject.toml
[tool.poetry]
name = "users-core"
version = "0.1.2"
description = ""
authors = ["kozlovmv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.8.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.5"
pre-commit = "^4.2.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"

[tool.ruff]
extend-select = ["I"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
